Fix wrong type of argument for PIL.image functions

+++ klippy/extras/display/st7789v.py
@@ -166,7 +166,7 @@ class DisplayBase:
                 # Convert the image to an array of bytes.
                 data = numpy.fromstring(strip.tobytes(), dtype=numpy.uint8)
                 # Convert from RGB888 (24-bit) to RGB565 (16-bit).
-                data565 = numpy.zeros((data.shape[0] / 3,), dtype=numpy.uint16)
+                data565 = numpy.zeros((int(data.shape[0] / 3),), dtype=numpy.uint16)
                 data565[:] += \
                     ((data[0::3] >> 3) & 0x1f).astype(numpy.uint16) << 11
                 data565[:] += \
